Separate data loading and data storage in BoondollarPriceManager
