+++ src/main/java/com/mraof/minestuck/entity/consort/ConsortRewardHandler.java
@@ -1,6 +1,6 @@
 package com.mraof.minestuck.entity.consort;
 
-import com.mraof.minestuck.util.BoondollarPriceManager;
+import com.mraof.minestuck.util.BoondollarPrices;
 import net.minecraft.resources.ResourceLocation;
 import net.minecraft.server.level.ServerLevel;
 import net.minecraft.util.RandomSource;
@@ -40,7 +40,7 @@ public class ConsortRewardHandler
 				}
 			}
 			
-			Optional<Integer> price = BoondollarPriceManager.getInstance().findPrice(stack, rand);
+			Optional<Integer> price = BoondollarPrices.getInstance().findPrice(stack, rand);
 			if (price.isPresent() && itemPriceList.size() < 9)
 				itemPriceList.add(Pair.of(stack, price.get()));
 			if (!price.isPresent())
